svg/painting/parsing/color-rendering-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing color-rendering with invalid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#ColorRenderingProperty"/>
    <h:meta name="assert" content="color-rendering supports only the grammar 'auto | optimizeQuality | optimizeSpeed'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_invalid_value("color-rendering", "optimizelegibility");
test_invalid_value("color-rendering", "auto optimizequality");

  ]]></script>
</svg>
